Add changes in Step 2, display random greeting messages using async fetch.
