+++ portfolio/src/main/webapp/index.html
@@ -9,10 +9,10 @@
 	<script src="script.js"></script>
 </head>
 
-<body>
+<body onload="getRandomQuoteUsingAsyncAwait()">
 	<div id="content">
 		<h1>About Myself</h1>
-		<p>Hi, welcome to Wenxuan's portfolio site. Glad that you are here:)</p>
+		<p id="greet">Hi, welcome to Wenxuan's portfolio site. Glad that you are here:)</p>
 		<p>The reason that I'm going with this background is because after trying so many options, this is the only
 			choice that doesn't block the texts and doesn't look weird.</p>
 		<p>I have chosen JetBrains Mono as the choice of font for this website. This is my favourite font so far, so I
